Complete

+++ src/com/company/Bat.java
@@ -2,7 +2,7 @@ package com.company;
 
 public class Bat extends Mammals implements canFly {
 
-    public Bat(double weight) {
+    public Bat(float weight) {
         super(weight);
     }
 }
