+++ src/com/company/Lion.java
@@ -2,7 +2,7 @@ package com.company;
 
 public class Lion extends Mammals {
 
-    public Lion(double weight) {
+    public Lion(float weight) {
         super(weight);
     }
