+++ src/com/company/Rat.java
@@ -2,7 +2,7 @@ package com.company;
 
 public class Rat extends Mammals implements canSwim {
 
-    public Rat(double weight) {
+    public Rat(float weight) {
         super(weight);
     }
 }
